fix interface implementation some little refactoring
