+++ pkg/bpmn_engine/engine.go
@@ -1,9 +1,6 @@
 package bpmn_engine
 
 import (
-	"crypto/md5"
-	"encoding/hex"
-	"encoding/xml"
 	"errors"
 	"fmt"
 	"github.com/bwmarrin/snowflake"
@@ -12,7 +9,6 @@ import (
 	"github.com/nitram509/lib-bpmn-engine/pkg/spec/BPMN20/activity"
 	"github.com/nitram509/lib-bpmn-engine/pkg/spec/BPMN20/process_instance"
 	"hash/adler32"
-	"io/ioutil"
 	"os"
 	"time"
 )
@@ -20,9 +16,9 @@ import (
 type BpmnEngine interface {
 	LoadFromFile(filename string) (*ProcessInfo, error)
 	LoadFromBytes(xmlData []byte) (*ProcessInfo, error)
-	AddTaskHandler(taskType string, handler func(job ActivatedJob))
-	CreateInstance(processKey int64, variableContext map[string]string) (*ProcessInstanceInfo, error)
-	CreateAndRunInstance(processKey int64, variableContext map[string]string) (*ProcessInstanceInfo, error)
+	AddTaskHandler(taskId string, handler func(job ActivatedJob))
+	CreateInstance(processKey int64, variableContext map[string]interface{}) (*ProcessInstanceInfo, error)
+	CreateAndRunInstance(processKey int64, variableContext map[string]interface{}) (*ProcessInstanceInfo, error)
 	RunOrContinueInstance(processInstanceKey int64) (*ProcessInstanceInfo, error)
 	GetName() string
 	GetProcessInstances() []*ProcessInstanceInfo
@@ -238,53 +234,6 @@ func checkOnlyOneAssociationOrPanic(event *BPMN20.TIntermediateCatchEvent) {
 		panic(fmt.Sprintf("Element with id=%s has %d incoming associations, but only 1 is supported by this engine.",
 			event.Id, len(event.IncomingAssociation)))
 	}
-}
-
-// LoadFromFile loads a given BPMN file by filename into the engine
-// and returns ProcessInfo details for the deployed workflow
-func (state *BpmnEngineState) LoadFromFile(filename string) (*ProcessInfo, error) {
-	xmlData, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return nil, err
-	}
-	return state.load(xmlData, filename)
-}
-
-// LoadFromBytes loads a given BPMN file by xmlData byte array into the engine
-// and returns ProcessInfo details for the deployed workflow
-func (state *BpmnEngineState) LoadFromBytes(xmlData []byte) (*ProcessInfo, error) {
-	return state.load(xmlData, "")
-}
-
-func (state *BpmnEngineState) load(xmlData []byte, resourceName string) (*ProcessInfo, error) {
-	md5sum := md5.Sum(xmlData)
-	var definitions BPMN20.TDefinitions
-	err := xml.Unmarshal(xmlData, &definitions)
-	if err != nil {
-		return nil, err
-	}
-
-	processInfo := ProcessInfo{
-		Version:     1,
-		definitions: definitions,
-	}
-	for _, process := range state.processes {
-		if process.BpmnProcessId == definitions.Process.Id {
-			if areEqual(process.checksumBytes, md5sum) {
-				return &process, nil
-			} else {
-				processInfo.Version = process.Version + 1
-			}
-		}
-	}
-	processInfo.BpmnProcessId = definitions.Process.Id
-	processInfo.ProcessKey = state.generateKey()
-	processInfo.checksumBytes = md5sum
-	state.processes = append(state.processes, processInfo)
-
-	state.exportProcessEvent(processInfo, xmlData, resourceName, hex.EncodeToString(md5sum[:]))
-
-	return &processInfo, nil
 }
 
 // AddTaskHandler registers a handler function to be called for service tasks with a given taskId
